requirements.txt
fastapi==0.100.0
uvicorn==0.24.0
sqlmodel==0.0.12
python-jose[cryptography]==3.3.0
bcrypt>=3.1.0
python-multipart==0.0.6
python-dotenv==1.0.0
email-validator==2.1.0